# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Motor Clínico - API para clínica médica integrativa
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /anamnese:
    get:
      operationId: listarAnamneses
      tags: [anamnese]
      summary: Listar anamneses
      parameters:
        - name: pacienteId
          in: query
          schema:
            type: integer
        - name: status
          in: query
          schema:
            type: string
            enum: [pendente, em_andamento, concluida, validada]
      responses:
        "200":
          description: Lista de anamneses
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Anamnese"
    post:
      operationId: criarAnamnese
      tags: [anamnese]
      summary: Criar nova anamnese
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CriarAnamneseBody"
      responses:
        "201":
          description: Anamnese criada
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Anamnese"
  /anamnese/{id}:
    get:
      operationId: obterAnamnese
      tags: [anamnese]
      summary: Obter anamnese por ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Anamnese encontrada
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Anamnese"
    put:
      operationId: atualizarAnamnese
      tags: [anamnese]
      summary: Atualizar respostas da anamnese
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AtualizarAnamneseBody"
      responses:
        "200":
          description: Anamnese atualizada
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Anamnese"
  /anamnese/{id}/ativar-motor:
    post:
      operationId: ativarMotorClinico
      tags: [anamnese, motor-clinico]
      summary: Ativar motor clínico com base na anamnese
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Sugestões geradas pelo motor clínico
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ResultadoMotorClinico"

  # Motor Clínico
...
  /filas:
    get:
      operationId: obterFilasOperacionais
      tags: [filas]
      summary: Obter filas operacionais por unidade
      parameters:
        - name: unidadeId
          in: query
          schema:
            type: integer
        - name: tipo
          in: query
          schema:
            type: string
            enum: [anamnese, validacao, procedimento, followup, pagamento]
      responses:
        "200":
          description: Filas operacionais
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FilasOperacionais"
  /filas/{itemId}/mover:
    post:
      operationId: moverItemFila
      tags: [filas]
      summary: Mover item na fila
      parameters:
        - name: itemId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MoverFilaBody"
      responses:
        "200":
          description: Item movido
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ItemFila"

  # Follow-up
...
  /dashboard/resumo:
    get:
      operationId: obterResumoDashboard
      tags: [dashboard]
      summary: Resumo geral do dashboard
      parameters:
        - name: unidadeId
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: Resumo do dashboard
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ResumoDashboard"
  /dashboard/metricas-motor:
    get:
      operationId: obterMetricasMotor
      tags: [dashboard]
      summary: Métricas do motor clínico
      parameters:
        - name: unidadeId
          in: query
          schema:
            type: integer
        - name: periodo
          in: query
          schema:
            type: string
            enum: [hoje, semana, mes]
            default: hoje
      responses:
        "200":
          description: Métricas do motor
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MetricasMotor"
  /dashboard/atividade-recente:
    get:
      operationId: obterAtividadeRecente
      tags: [dashboard]
      summary: Atividade recente do sistema
      parameters:
        - name: limite
          in: query
          schema:
            type: integer
            default: 20
        - name: unidadeId
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: Lista de atividades recentes
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AtividadeRecente"
  /dashboard/filas-resumo:
    get:
      operationId: obterResumoFilas
      tags: [dashboard]
      summary: Resumo rápido de todas as filas
      parameters:
        - name: unidadeId
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: Resumo das filas
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ResumoFilas"
components:
  schemas:
...
    Anamnese:
      type: object
      properties:
        id:
          type: integer
        pacienteId:
          type: integer
        pacienteNome:
          type: string
        status:
          type: string
          enum: [pendente, em_andamento, concluida, validada]
        respostasClincias:
          type: object
          additionalProperties: true
        respostasFinanceiras:
          type: object
          additionalProperties: true
        respostasPreferencias:
          type: object
          additionalProperties: true
        sinaisSemanticos:
          type: array
          items:
            type: string
        motorAtivadoEm:
          type: string
          format: date-time
        criadoEm:
          type: string
          format: date-time
        atualizadoEm:
          type: string
          format: date-time
      required:
        - id
        - pacienteId
        - status
        - criadoEm
        - atualizadoEm
    CriarAnamneseBody:
      type: object
      properties:
        pacienteId:
          type: integer
      required:
        - pacienteId
    AtualizarAnamneseBody:
      type: object
      properties:
        respostasClincias:
          type: object
          additionalProperties: true
        respostasFinanceiras:
          type: object
          additionalProperties: true
        respostasPreferencias:
          type: object
          additionalProperties: true
        status:
          type: string
          enum: [pendente, em_andamento, concluida, validada]
    ResultadoMotorClinico:
      type: object
      properties:
        anamneseId:
          type: integer
        sinaisSemanticos:
          type: array
          items:
            type: string
        sugestoesGeradas:
          type: array
          items:
            $ref: "#/components/schemas/SugestaoClinica"
        totalSugestoes:
          type: integer
      required:
        - anamneseId
        - sinaisSemanticos
        - sugestoesGeradas
        - totalSugestoes
    SugestaoClinica:
      type: object
      properties:
        id:
          type: integer
        anamneseId:
          type: integer
        pacienteId:
          type: integer
        pacienteNome:
          type: string
        tipo:
          type: string
          enum: [exame, formula, injetavel_im, injetavel_ev, implante, protocolo]
        itemTerapeuticoId:
          type: integer
        itemNome:
          type: string
        itemDescricao:
          type: string
        justificativa:
          type: string
        prioridade:
          type: string
          enum: [baixa, media, alta, urgente]
        status:
          type: string
          enum: [pendente, validado, rejeitado, em_execucao, concluido]
        validadoPorId:
          type: integer
        validadoPorNome:
          type: string
        validadoEm:
          type: string
          format: date-time
        observacaoValidacao:
          type: string
        criadoEm:
          type: string
          format: date-time
      required:
        - id
        - anamneseId
        - pacienteId
        - tipo
        - itemNome
        - prioridade
        - status
        - criadoEm
...
    FilasOperacionais:
      type: object
      properties:
        filaAnamnese:
          type: array
          items:
            $ref: "#/components/schemas/ItemFila"
        filaValidacao:
          type: array
          items:
            $ref: "#/components/schemas/ItemFila"
        filaProcedimento:
          type: array
          items:
            $ref: "#/components/schemas/ItemFila"
        filaFollowup:
          type: array
          items:
            $ref: "#/components/schemas/ItemFila"
        filaPagamento:
          type: array
          items:
            $ref: "#/components/schemas/ItemFila"
      required:
        - filaAnamnese
        - filaValidacao
        - filaProcedimento
        - filaFollowup
        - filaPagamento
    ItemFila:
      type: object
      properties:
        id:
          type: integer
        tipo:
          type: string
          enum: [anamnese, validacao, procedimento, followup, pagamento]
        pacienteId:
          type: integer
        pacienteNome:
          type: string
        descricao:
          type: string
        prioridade:
          type: string
          enum: [baixa, media, alta, urgente]
        status:
          type: string
        responsavelId:
          type: integer
        responsavelNome:
          type: string
        unidadeId:
          type: integer
        criadoEm:
          type: string
          format: date-time
        prazo:
          type: string
          format: date-time
      required:
        - id
        - tipo
        - pacienteId
        - pacienteNome
        - prioridade
        - status
        - unidadeId
        - criadoEm
    MoverFilaBody:
      type: object
      properties:
        novoStatus:
          type: string
        responsavelId:
          type: integer
        observacao:
          type: string
      required:
        - novoStatus
...
    ResumoDashboard:
      type: object
      properties:
        totalPacientes:
          type: integer
        anamnesesPendentes:
          type: integer
        sugestoesPendentesValidacao:
          type: integer
        procedimentosHoje:
          type: integer
        followupAtrasados:
          type: integer
        pagamentosPendentes:
          type: integer
        receitaHoje:
          type: number
        unidadeNome:
          type: string
      required:
        - totalPacientes
        - anamnesesPendentes
        - sugestoesPendentesValidacao
        - procedimentosHoje
        - followupAtrasados
        - pagamentosPendentes
        - receitaHoje
    MetricasMotor:
      type: object
      properties:
        totalSugestoes:
          type: integer
        sugestoesValidadas:
          type: integer
        sugestoesRejeitadas:
          type: integer
        taxaValidacao:
          type: number
        sugestoesPorTipo:
          type: array
          items:
            type: object
            properties:
              tipo:
                type: string
              total:
                type: integer
              validadas:
                type: integer
        periodo:
          type: string
      required:
        - totalSugestoes
        - sugestoesValidadas
        - sugestoesRejeitadas
        - taxaValidacao
        - sugestoesPorTipo
        - periodo
    AtividadeRecente:
      type: object
      properties:
        id:
          type: integer
        tipo:
          type: string
          enum: [anamnese, validacao, pagamento, followup, cadastro, procedimento]
        descricao:
          type: string
        pacienteNome:
          type: string
        usuarioNome:
          type: string
        criadoEm:
          type: string
          format: date-time
      required:
        - id
        - tipo
        - descricao
        - criadoEm
    ResumoFilas:
      type: object
      properties:
        anamnese:
          type: integer
        validacao:
          type: integer
        procedimento:
          type: integer
        followup:
          type: integer
        pagamento:
          type: integer
        totalUrgente:
          type: integer
      required:
        - anamnese
        - validacao
        - procedimento
        - followup
        - pagamento
        - totalUrgente